+++ enigma/cogs/basics.py
@@ -159,7 +159,7 @@ class Basics(Cog):
             name='Support',
             value='If you\'d like to help to develop this bot please check GitHub. (Link above)\n'
                   'And if **you** need support please check'
-                  ' [support server](https://discord.gg/SRdmrPpf2z "3N1GMA Support Server").',
+                  ' [official support server](https://discord.gg/SRdmrPpf2z "3N1GMA Support Server").',
             inline=False
         ).add_field(
             name='More about bot',